Move nested_list, byte_string, as_num256 syntax tests to own files.
