+++ tests/parser/syntax/test_len.py
@@ -1,0 +1,41 @@
+import pytest
+from pytest import raises
+
+from viper import compiler
+from viper.exceptions import TypeMismatchException
+
+
+fail_list = [
+    """
+def foo(inp: num) -> num:
+    return len(inp)
+    """,
+    """
+def foo(inp: num) -> address:
+    return len(inp)
+    """
+]
+
+
+@pytest.mark.parametrize('bad_code', fail_list)
+def test_block_fail(bad_code):
+
+    if isinstance(bad_code, tuple):
+        with raises(bad_code[1]):
+            compiler.compile(bad_code[0])
+    else:
+        with raises(TypeMismatchException):
+            compiler.compile(bad_code)
+
+
+valid_list = [
+    """
+def foo(inp: bytes <= 10) -> num:
+    return len(inp)
+    """
+]
+
+
+@pytest.mark.parametrize('good_code', valid_list)
+def test_list_success(good_code):
+    assert compiler.compile(good_code) is not None
