+++ tests/test_invalids.py
@@ -18,6 +18,7 @@ def must_succeed(code):
 
 # TEST CASES
 
+
 must_succeed("""
 x: num[3]
 """)
@@ -73,7 +74,6 @@ def foo():
     self.b = 7.5
 """)
 
-
 must_succeed("""
 b: decimal
 def foo():
@@ -85,61 +85,6 @@ b: num[5]
 def foo():
     self.b = 7
 """, TypeMismatchException)
-
-must_succeed("""
-b: decimal[5]
-def foo():
-    self.b[0] = 7
-""")
-
-must_fail("""
-b: num[5]
-def foo():
-    self.b[0] = 7.5
-""", TypeMismatchException)
-
-must_succeed("""
-b: num[5]
-def foo():
-    a: num[5]
-    self.b[0] = a[0]
-""")
-
-must_fail("""
-b: num[5]
-def foo():
-    x = self.b[0][1]
-""", TypeMismatchException)
-
-must_fail("""
-b: num[5]
-def foo():
-    x = self.b[0].cow
-""", TypeMismatchException)
-
-must_fail("""
-b: {foo: num}
-def foo():
-    self.b = {foo: 1, foo: 2}
-""", TypeMismatchException)
-
-must_fail("""
-b: {foo: num, bar: num}
-def foo():
-    x = self.b.cow
-""", TypeMismatchException)
-
-must_fail("""
-b: {foo: num, bar: num}
-def foo():
-    x = self.b[0]
-""", TypeMismatchException)
-
-must_succeed("""
-b: {foo: num, bar: num}
-def foo():
-    x = self.b.bar
-""")
 
 must_succeed("""
 b: num[num]
@@ -177,7 +122,6 @@ def foo():
     self.b[-3] = 5
 """)
 
-
 must_succeed("""
 def foo():
     x: num[5]
@@ -209,66 +153,6 @@ def foo():
     self.foo[0] = 5
 """)
 
-must_succeed("""
-foo: num[3]
-def foo():
-    self.foo = [1, 2, 3]
-""")
-
-must_fail("""
-foo: num[3]
-def foo():
-    self.foo = [1, 2, 3, 4]
-""", TypeMismatchException)
-
-must_fail("""
-foo: num[3]
-def foo():
-    self.foo = [1, 2]
-""", TypeMismatchException)
-
-must_fail("""
-foo: num[3]
-def foo():
-    self.foo = [1, [2], 3]
-""", TypeMismatchException)
-
-must_fail("""
-bar: num[3][3]
-def foo():
-    self.bar = 5
-""", TypeMismatchException)
-
-must_fail("""
-bar: num[3][3]
-def foo():
-    self.bar = [2, 5]
-""", TypeMismatchException)
-
-must_fail("""
-bar: num[3][3]
-def foo():
-    self.bar = [[1, 2], [3, 4, 5], [6, 7, 8]]
-""", TypeMismatchException)
-
-must_succeed("""
-bar: num[3][3]
-def foo():
-    self.bar = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
-""")
-
-must_fail("""
-bar: num[3][3]
-def foo():
-    self.bar = [[1, 2, 3], [4, 5, 6], [7, 8, 9.0]]
-""", TypeMismatchException)
-
-must_succeed("""
-bar: decimal[3][3]
-def foo():
-    self.bar = [[1, 2, 3], [4, 5, 6], [7, 8, 9.0]]
-""")
-
 must_fail("""
 def foo() -> address:
     return as_unitless_number([1, 2, 3])
@@ -284,52 +168,6 @@ def baa() -> decimal:
     return 2.0**2
 """, TypeMismatchException)
 
-must_fail("""
-def foo(inp: bytes <= 10) -> bytes <= 3:
-    return slice(inp, start=4.0, len=3)
-""", TypeMismatchException)
-
-must_succeed("""
-def foo(inp: bytes <= 10) -> num:
-    return len(inp)
-""")
-
-must_fail("""
-def foo(inp: num) -> num:
-    return len(inp)
-""", TypeMismatchException)
-
-must_succeed("""
-def foo() -> bytes <= 10:
-    return "badminton"
-""")
-
-must_fail("""
-def foo() -> bytes <= 10:
-    return "badmintonzz"
-""", TypeMismatchException)
-
-must_succeed("""
-def foo() -> bytes <= 10:
-    return slice("badmintonzzz", start=1, len=10)
-""")
-
-must_fail("""
-def foo() -> bytes <= 10:
-    x = '0x1234567890123456789012345678901234567890'
-    x = 0x1234567890123456789012345678901234567890
-""", TypeMismatchException)
-
-must_succeed("""
-def foo():
-    x = "¡très bien!"
-""")
-
-must_succeed("""
-def convert1(inp: bytes32) -> num256:
-    return as_num256(inp)
-""")
-
 must_succeed("""
 def foo():
     throw
@@ -343,12 +181,6 @@ def goo():
     self.foo()
 """)
 
-must_fail("""
-x: {cow: num, cor: num}
-def foo():
-    self.x.cof = 1
-""", TypeMismatchException)
-
 must_succeed("""
 def foo():
     MOOSE = 45
@@ -357,11 +189,6 @@ def foo():
 must_fail("""
 def foo():
     x = -self
-""", TypeMismatchException)
-
-must_fail("""
-def foo() -> num:
-    return {cow: 5, dog: 7}
 """, TypeMismatchException)
 
 must_fail("""
